# app/Http/Resources/VehicleMakeResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

class VehicleMakeResource extends JsonResource
{
    public function toArray(Request $request): array
    {
        return [
            'id' => $this->id,
            'name' => $this->name,
            'country_of_origin' => $this->country_of_origin,
            'founded_year' => $this->founded_year,
        ];
    }
}