# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vehicle Management API",
        "description": "API documentation for Vehicle Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        },
        {
            "url": "http://localhost/api",
            "description": "Docker API Server"
        }
    ],
    "paths": {
...
        "/vehicles/{vehicle}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get detailed information about a specific vehicle",
                "operationId": "5d8cf15c75741577ed197fba4ec26262",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles/{vehicle}/parameters": {
            "patch": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Update a specific parameter of a vehicle",
                "operationId": "606c5eec8cc48ff2a20ba40274a47dad",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parameter",
                                    "value"
                                ],
                                "properties": {
                                    "parameter": {
                                        "type": "string",
                                        "example": "top_speed"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "220"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle parameter updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "model": {
                                            "type": "string",
                                            "example": "Camry"
                                        },
                                        "year": {
                                            "type": "integer",
                                            "example": 2023
                                        },
                                        "attributes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "top_speed"
                                                    },
                                                    "value": {
                                                        "type": "string",
                                                        "example": "220"
                                                    },
                                                    "unit": {
                                                        "type": "string",
                                                        "example": "km/h"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicle-types/{vehicleType}/makes": {
            "get": {
                "tags": [
                    "Vehicle Makes"
                ],
                "summary": "Get all vehicle makes that manufacture a specific type of vehicle",
                "operationId": "136b72451789fe56f7a307fa77e915ae",
                "parameters": [
                    {
                        "name": "vehicleType",
                        "in": "path",
                        "description": "Vehicle Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of vehicle makes retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VehicleMakeResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle type not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "VehicleResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "model": {
                        "type": "string",
                        "example": "Camry"
                    },
                    "year": {
                        "type": "integer",
                        "example": 2023
                    },
                    "type": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Sedan"
                            }
                        },
                        "type": "object"
                    },
                    "make": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Toyota"
                            }
                        },
                        "type": "object"
                    },
                    "attributes": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "top_speed"
                                },
                                "value": {
                                    "type": "string",
                                    "example": "210"
                                },
                                "unit": {
                                    "type": "string",
                                    "example": "km/h"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "VehicleMakeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "country_of_origin": {
                        "type": "string",
                        "example": "Japan"
                    },
                    "founded_year": {
                        "type": "integer",
                        "example": 1937
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "properties": {
                            "field": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Error message for field"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}